# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.AddTorrentRequest:
    properties:
      magnetUri:
        type: string
      torrentFile:
        description: base64 encoded torrent file content
        type: string
    type: object
...
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.BufferInfo:
    properties:
      buffer_end:
        description: Buffer end
        type: integer
      buffer_size:
        description: Buffer size
        type: integer
      buffer_start:
        description: Buffer start
        type: integer
      buffered_bytes:
        description: Bytes buffered
        type: integer
      buffered_percent:
        description: Buffer load percentage
        type: number
      current_position:
        description: Current position in bytes
        type: integer
      download_speed:
        description: Download speed (bytes/sec)
        type: integer
      file_index:
        type: integer
      is_buffering:
        description: Is buffering in progress
        type: boolean
      torrent_id:
        type: string
    type: object
...
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.FileInfo:
    properties:
      index:
        type: integer
      name:
        type: string
      size:
        type: integer
    type: object
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.FileListResponse:
    properties:
      files:
        items:
          $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.FileInfo'
        type: array
      hasMore:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      totalCount:
        type: integer
    type: object
...
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.SelectFileRequest:
    properties:
      fileIndex:
        type: integer
    type: object
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.SetBufferPositionRequest:
    properties:
      position:
        description: Position in bytes
        type: integer
    type: object
...
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.StreamTicketResponse:
    properties:
      ticket:
        type: string
      torrentId:
        type: string
    type: object
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.TorrentInfo:
    properties:
      id:
        type: string
      name:
        type: string
      progress:
        type: number
      size:
        type: integer
      status:
        type: string
    type: object
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.TorrentListResponse:
    properties:
      hasMore:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      torrents:
        items:
          $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.TorrentInfo'
        type: array
      totalCount:
        type: integer
    type: object
...
host: localhost:8889
info:
  contact:
    email: support@torrsyncplayer.local
    name: TorrSyncPlayer Support
    url: https://github.com/blagovibe/TorrSyncPlayer/issues
  description: HTTP API for TorrSyncPlayer — torrent player with P2P playback synchronization.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/blagovibe/TorrSyncPlayer
  title: TorrSyncPlayer API
  version: "1.0"
paths:
...
  /api/v1/torrents:
    get:
      description: Returns a list of all torrents with pagination
      parameters:
      - default: 20
        description: Record limit
        in: query
        maximum: 100
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.TorrentListResponse'
      summary: List torrents
      tags:
      - torrents
    post:
      consumes:
      - application/json
      description: Adds a torrent via magnet link or torrent file and returns its
        information
      parameters:
      - description: Magnet URI or Torrent File (base64)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.AddTorrentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.TorrentInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
      summary: Add torrent
      tags:
      - torrents
  /api/v1/torrents/{id}:
    delete:
      description: Removes a torrent by its ID
      parameters:
      - description: Torrent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
      summary: Remove torrent
      tags:
      - torrents
  /api/v1/torrents/{id}/buffer/info:
    get:
      description: Returns buffer state information for the torrent
      parameters:
      - description: Torrent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.BufferInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
      summary: Buffer info
      tags:
      - torrents
  /api/v1/torrents/{id}/buffer/position:
    post:
      consumes:
      - application/json
      description: Sets the current playback position for buffer optimization
      parameters:
      - description: Torrent ID
        in: path
        name: id
        required: true
        type: string
      - description: Position in bytes
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.SetBufferPositionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
      summary: Set buffer position
      tags:
      - torrents
  /api/v1/torrents/{id}/files:
    get:
      description: Returns a list of torrent files with pagination
      parameters:
      - description: Torrent ID
        in: path
        name: id
        required: true
        type: string
      - default: 20
        description: Record limit
        in: query
        maximum: 100
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.FileListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
      summary: List torrent files
      tags:
      - torrents
  /api/v1/torrents/{id}/select:
    post:
      consumes:
      - application/json
      description: Selects a torrent file for streaming by index
      parameters:
      - description: Torrent ID
        in: path
        name: id
        required: true
        type: string
      - description: File index
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.SelectFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
      summary: Select file for streaming
      tags:
      - torrents
  /api/v1/torrents/{id}/stream:
    get:
      description: Streams the selected torrent file with Range request support
      parameters:
      - description: Torrent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
      summary: Stream file
      tags:
      - torrents
  /api/v1/torrents/{id}/stream-ticket:
    post:
      description: Returns a signed ticket authorizing streaming of the given torrent
      parameters:
      - description: Torrent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.StreamTicketResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
      summary: Issue stream ticket
      tags:
      - torrents
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"